output: Extract candidate preview snippets
Candidate preview summary derived target summaries and also owned snippet
line formatting. The renderer and diff renderer reached through the summary
module for gutters, width calculation, and range highlighting.

That made target-summary derivation a catch-all dependency for visual
snippet primitives. Moving those primitives gives summary and rendering
code separate module names.

This commit moves snippet records and formatting helpers into
output.candidate_preview_snippets and updates preview renderers to import
that module directly.

Subsequent commits will add architecture coverage for the snippet boundary.
